changed rating columns types, removed Autowired annotation

+++ src/main/java/hillel/spring/doctor/dto/ReviewInputDto.java
@@ -11,9 +11,9 @@ public class ReviewInputDto {
     private Integer doctorRecordId;
 
     private Integer serviceRating;
-    private Byte equipmentRating;
-    private Byte qualificationRating;
-    private Byte treatmentRating;
-    private Byte totalRating;
+    private Integer equipmentRating;
+    private Integer qualificationRating;
+    private Integer treatmentRating;
+    private Integer totalRating;
     private String reviewComment;
 }
